add forever running mode
